CREATE FILTER COMPONENT
Filters based on type of exercise and sort exercise by time of creation or popularity.
Add field "count" denoting the number of people who joined the exercise to the exercise
component in order to sort by popularity. Sort by time of creation was based on the
assumption that exercises with a later creation time will have a larger id. May need to
refactor after connecting to db. Need to add feature where count increments as users
join the challenge.