ADD SIDEBAR THAT SHOWS EXERCISES JOINED
Add functionality to JOIN button in Exercises.js so that when clicked, the number of
people joined (count field of exercise) for the corresponding exercise gets incremented
by 1. No incrementation happens for already joined exercises. The sidebar shows total
number of exercises joined and displays preview of exercise (image and title). Includes
remove button for user to withdraw from exercise.